fix(state): use a signed send quota so retransmit cannot underflow
sendQuota.quota and initialQuota were uint16, but Retransmit (noWait) decrements the
quota without blocking. On a reconnect that retransmits more stored messages than the
receive maximum, the quota is meant to go negative (as the existing comments note); with
uint16 it instead wrapped to ~65535, which:

  - makes Acquire stop blocking (quota is always > 0), so the broker's receive maximum is
    no longer honoured, and
  - breaks Release (s.quota < initialQuota becomes false), which then returns
    ErrUnexpectedRelease and never recovers.

Make quota and initialQuota signed (int) so the counter can legitimately go negative,
matching the intended design.

+++ paho/session/state/sendquota.go
@@ -46,16 +46,21 @@ var ErrUnexpectedRelease = errors.New("release called when quota at initial valu
 
 // newSendQuota creates a new tracker limited to quota concurrent messages
 func newSendQuota(quota uint16) *sendQuota {
-	w := &sendQuota{initialQuota: quota, quota: quota}
+	w := &sendQuota{initialQuota: int(quota), quota: int(quota)}
 	return w
 }
 
 // sendQuota provides a way to bound concurrent access to a resource.
 // The callers can request access with a given weight.
+//
+// quota is a signed counter: Retransmit (noWait) decrements it without blocking, so on a
+// reconnect that retransmits more stored messages than the receive maximum it can legitimately
+// go negative (over-allocated). It must therefore be a signed type — a uint16 here would wrap to
+// a large value, permanently defeating the receive-maximum bound and breaking Release.
 type sendQuota struct {
 	mu           sync.Mutex
-	initialQuota uint16
-	quota        uint16
+	initialQuota int
+	quota        int
 	waiters      []chan<- struct{} // using a slice because would generally expect this to be small
 }
